サイト基本設定を保存できるように調整した

+++ lib/Baser/Controller/SiteConfigsController.php
@@ -288,7 +288,7 @@ class SiteConfigsController extends AppController {
 		$data['SiteConfig']['smart_url'] = $this->BcManager->smartUrl();
 		$data['SiteConfig']['site_url'] = Configure::read('BcEnv.siteUrl');
 		$data['SiteConfig']['ssl_url'] = Configure::read('BcEnv.sslUrl');
-		$data['SiteConfig']['admin_ssl'] = Configure::read('BcApp.adminSsl');
+		$data['SiteConfig']['admin_ssl'] = (int) Configure::read('BcApp.adminSsl');
 		$data['SiteConfig']['mobile'] = Configure::read('BcApp.mobile');
 		$data['SiteConfig']['smartphone'] = Configure::read('BcApp.smartphone');
 		if(is_null($data['SiteConfig']['mobile'])) {
